Use pundit for authorisation in gql

+++ app/graphql/types/mutation_type.rb
@@ -1,17 +1,13 @@
-CommentInputType = GraphQL::InputObjectType.define do
-  name "CommentInputType"
-
-  argument :productRef, !types.String
-  argument :content, !types.String
-end
-
 Types::MutationType = GraphQL::ObjectType.define do
   name "Mutation"
 
   field :toggleLike, Types::ProductType do
     argument :productRef, !types.String
-    resolve ->(_obj, args, ctx) {
-      @product = OpenStruct.new(product_ref: args[:productRef])
+    resolve Resolver.new ->(obj, args, ctx) {
+      use(Plugins::Pundit, obj: obj, args: args, ctx: ctx)
+
+      @product = Product.new(product_ref: args[:productRef])
+      authorize(@product)
       like = Like.find_by(product_ref: args[:productRef], customer_ref: ctx[:current_user].customer_ref)
       if like
         like.destroy!
@@ -23,17 +19,26 @@ Types::MutationType = GraphQL::ObjectType.define do
   end
 
   field :addComment, Types::CommentType do
-    argument :comment, !CommentInputType
+    argument :comment, !Types::CommentInputType
 
-    resolve ->(_obj, args, ctx) {
-      Comment.create!(args[:comment].to_h.transform_keys { |key| key.to_s.underscore }.merge(user: ctx[:current_user]))
+    resolve Resolver.new ->(obj, args, ctx) {
+      use(Plugins::Pundit, obj: obj, args: args, ctx: ctx)
+
+      comment_args = permitted_attributes(Comment).merge(user: current_user)
+      @comment = policy_scope(Comment).new(comment_args)
+      authorize(@comment)
+      @comment.save!
+      @comment
     }
   end
 
   field :toggleUpvote, Types::CommentType do
     argument :commentId, !types.ID
-    resolve ->(_obj, args, ctx) {
+    resolve Resolver.new ->(obj, args, ctx) {
+      use(Plugins::Pundit, obj: obj, args: args, ctx: ctx)
+
       @comment = Comment.find(args[:commentId])
+      authorize(@comment)
       upvote = Upvote.find_by(comment: @comment, user: ctx[:current_user])
       if upvote
         upvote.destroy!
@@ -46,8 +51,11 @@ Types::MutationType = GraphQL::ObjectType.define do
 
   field :flag, Types::CommentType do
     argument :commentId, !types.ID
-    resolve ->(_obj, args, ctx) {
+    resolve Resolver.new ->(obj, args, ctx) {
+      use(Plugins::Pundit, obj: obj, args: args, ctx: ctx)
+
       @comment = Comment.find(args[:commentId])
+      authorize(@comment)
       inappropriate_flag = InappropriateFlag.find_by(comment: @comment, user: ctx[:current_user])
       InappropriateFlag.create!(comment: @comment, user: ctx[:current_user]) unless inappropriate_flag
       @comment.reload
